Limit set of indicies queried on cloudelastic
Cloudelastic has two ways to target the indices to query. There are
the exact index names, such as foowiki_content_12345, and there are
aliases pointing to the current live index, such as foowiki_content.

The prior index selection method, taking everything with `*,*:*`,
captured not only the live indices but also indices that were not live
and not intended to be queried.

Patch adjusts the index selection method such that it will target the
known named aliases provided by CirrusSearch.